clean project files structure by moving files and adding proper references
