+++ src/main/java/com/capgemini/wsb/fitnesstracker/user/api/UserBasicEmailDto.java
@@ -1,4 +1,4 @@
-package com.capgemini.wsb.fitnesstracker.user.internal;
+package com.capgemini.wsb.fitnesstracker.user.api;
 
 import jakarta.annotation.Nullable;
