+++ src/main/java/com/capgemini/wsb/fitnesstracker/user/internal/UserDto.java
@@ -5,7 +5,7 @@ import jakarta.annotation.Nullable;
 
 import java.time.LocalDate;
 
-record UserDto(@Nullable Long Id, String firstName, String lastName,
+record serDto(@Nullable Long Id, String firstName, String lastName,
                @JsonFormat(pattern = "yyyy-MM-dd") LocalDate birthdate,
                String email) {
