+++ src/main/java/com/capgemini/wsb/fitnesstracker/user/internal/UserMapper.java
@@ -1,6 +1,9 @@
 package com.capgemini.wsb.fitnesstracker.user.internal;
 
 import com.capgemini.wsb.fitnesstracker.user.api.User;
+import com.capgemini.wsb.fitnesstracker.user.api.UserBasicDto;
+import com.capgemini.wsb.fitnesstracker.user.api.UserBasicEmailDto;
+import com.capgemini.wsb.fitnesstracker.user.api.UserDto;
 import org.springframework.stereotype.Component;
 
 @Component
